requirements.txt
pika==1.3.1
lxml>=5.0.0
python-dotenv==1.0.0
psycopg2-binary>=2.9.0

# Testing
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
